+++ gdscript/src/main/kotlin/gdscript/annotator/GdFlowStmtAnnotator.kt
@@ -1,0 +1,28 @@
+package gdscript.annotator
+
+import com.intellij.lang.annotation.AnnotationHolder
+import com.intellij.lang.annotation.Annotator
+import com.intellij.lang.annotation.HighlightSeverity
+import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
+import gdscript.psi.*
+
+class GdFlowStmtAnnotator : Annotator {
+
+    override fun annotate(element: PsiElement, holder: AnnotationHolder) {
+        if (element !is GdFlowSt) return
+        val txt = element.type
+        if (txt != "continue" && txt != "break") return
+
+        PsiTreeUtil.findFirstParent(element) {
+            it is GdMethodDeclTl || it is GdFuncDeclEx || it is GdForSt || it is GdWhileSt
+        }?.let {
+            if (it is GdMethodDeclTl || it is GdFuncDeclEx) {
+                holder
+                    .newAnnotation(HighlightSeverity.ERROR, "Cannot use '$txt' outside of a loop")
+                    .range(element.textRange)
+                    .create()
+            }
+        }
+    }
+}